// apps/central-front/src/app/core/model/global/confirm.dialog.class.ts
import {Observable} from 'rxjs';

/**
 * Input data for the {@link ConfirmDialogComponent}
 */
export interface ConfirmDialogInput {
  /**
   * Title of the dialog, supports HTML
   */
  title: string;

  /**
   * Content of the dialog, supports HTML
   */
  content: string;

  /**
   * Optional.If true the title and content are translated.
   */
  translateTitleAndContent?: boolean;

  /**
   * Optional observable, if filled, the observable is called
   * when the user press 'Yes' and the dialog is can't be closed until
   * the observable completes. The dialog returns the observable result
   */
  observable?: Observable<any>;

  /**
   * Optional message. If filled, a snackbar with the message is shown
   * when click 'yes' or when the observable completes successfully.
   */
  successMessage?: string;

  /**
   * Optional with the message. If the successMessage is set and the boolean
   * is set to true, the successMessage is translated.
   */
  translateMessage?: boolean;
}

/**
 * Object returned by the {@link ConfirmDialogResult}
 * when calling the openConfirmDialog from {@link DialogService}
 */
export interface ConfirmDialogResult<T = any> {
  /**
   * True if the user pressed 'Yes', false otherwise.
   */
  choice: boolean;

  /**
   * Fill by the observable result if the observable was provided
   */
  result?: T;
}